+++ src/network.rs
@@ -8,23 +8,23 @@ use helpers::{d_sigmoid, get_weight_delta, sigmoid};
 
 #[derive(Debug)]
 pub struct Network {
-    weights: Vec<Matrix<f32>>,
-    biases: Vec<Matrix<f32>>,
-    learning_rate: f32,
+    weights: Vec<Matrix<f64>>,
+    biases: Vec<Matrix<f64>>,
+    learning_rate: f64,
 }
 
 impl Network {
-    pub fn new(layers: Vec<usize>, number_of_inputs: usize, learning_rate: f32) -> Self {
+    pub fn new(layers: Vec<usize>, number_of_inputs: usize, learning_rate: f64) -> Self {
         let mut rng = rand::thread_rng();
-        let mut weights: Vec<Matrix<f32>> = Vec::with_capacity(layers.len());
-        let mut biases: Vec<Matrix<f32>> = Vec::with_capacity(layers.len());
+        let mut weights: Vec<Matrix<f64>> = Vec::with_capacity(layers.len());
+        let mut biases: Vec<Matrix<f64>> = Vec::with_capacity(layers.len());
         let mut prev_layer_outputs = number_of_inputs;
         for layer in layers {
             weights.push(Matrix::from_fn(layer, prev_layer_outputs, |_, _| {
-                rng.gen::<f32>() - rng.gen::<f32>()
+                rng.gen::<f64>() - rng.gen::<f64>()
             }));
             biases.push(Matrix::from_fn(layer, 1, |_, _| {
-                rng.gen::<f32>() - rng.gen::<f32>()
+                rng.gen::<f64>() - rng.gen::<f64>()
             }));
             prev_layer_outputs = layer;
         }
@@ -35,7 +35,7 @@ impl Network {
         }
     }
 
-    pub fn feed_forward(&self, inputs: Vec<f32>) -> Vec<f32> {
+    pub fn feed_forward(&self, inputs: Vec<f64>) -> Vec<f64> {
         if self.weights.is_empty() {
             panic!("No starting weights set!");
         }
@@ -46,7 +46,7 @@ impl Network {
             panic!("Inputs length needs to be {}", self.weights[0].cols());
         }
 
-        let mut layer_output: Matrix<f32> = Matrix::new(inputs.len(), 1, inputs);
+        let mut layer_output: Matrix<f64> = Matrix::new(inputs.len(), 1, inputs);
 
         for layer in 0..self.biases.len() {
             layer_output = Matrix::new(
@@ -56,7 +56,7 @@ impl Network {
                     .into_vec()
                     .iter()
                     .map(|x| sigmoid(*x))
-                    .collect::<Vec<f32>>(),
+                    .collect::<Vec<f64>>(),
             );
         }
         layer_output.into_vec()
@@ -65,7 +65,7 @@ impl Network {
     pub fn train(&mut self, training_data: Vec<TrainingData>) {
         for data in training_data.into_iter() {
             // feed-forward
-            let mut layer_output: Vec<Matrix<f32>> =
+            let mut layer_output: Vec<Matrix<f64>> =
                 vec![Matrix::new(data.inputs.len(), 1, data.inputs)];
             for layer in 0..self.biases.len() {
                 layer_output.push(Matrix::new(
@@ -75,13 +75,13 @@ impl Network {
                         .into_vec()
                         .iter()
                         .map(|x| sigmoid(*x))
-                        .collect::<Vec<f32>>(),
+                        .collect::<Vec<f64>>(),
                 ));
             }
 
             // back-propagate
-            let mut layer_gradients: Vec<Matrix<f32>> =
-                vec![Matrix::<f32>::zeros(1, 1); layer_output.len()];
+            let mut layer_gradients: Vec<Matrix<f64>> =
+                vec![Matrix::<f64>::zeros(1, 1); layer_output.len()];
             // last layer, aka output layer, gets special calculation
             let last_layer_index = layer_output.len() - 1;
             layer_gradients[last_layer_index] = Matrix::new(
@@ -94,7 +94,7 @@ impl Network {
                     .map(|(i, output)| {
                         (data.target[i] - *output) * d_sigmoid(*output) * self.learning_rate
                     })
-                    .collect::<Vec<f32>>(),
+                    .collect::<Vec<f64>>(),
             );
             // will need to skip layer_gradients[0] as that's the inputs
             for i in (1..last_layer_index).rev() {
@@ -104,8 +104,8 @@ impl Network {
                     (self.weights[i].transpose() * &layer_gradients[i + 1])
                         .into_vec()
                         .iter()
-                        .map(|x: &f32| d_sigmoid(*x) * self.learning_rate)
-                        .collect::<Vec<f32>>(),
+                        .map(|x: &f64| d_sigmoid(*x) * self.learning_rate)
+                        .collect::<Vec<f64>>(),
                 );
             }
 
@@ -147,9 +147,9 @@ impl Network {
         Ok(())
     }
 
-    pub fn from_data(data: NetworkData, learning_rate: f32) -> Self {
-        let mut weights: Vec<Matrix<f32>> = Vec::with_capacity(data.weights.len());
-        let mut biases: Vec<Matrix<f32>> = Vec::with_capacity(data.biases.len());
+    pub fn from_data(data: NetworkData, learning_rate: f64) -> Self {
+        let mut weights: Vec<Matrix<f64>> = Vec::with_capacity(data.weights.len());
+        let mut biases: Vec<Matrix<f64>> = Vec::with_capacity(data.biases.len());
         for layer in 0..data.weights.len() {
             weights.push(Matrix::new(
                 data.weights[layer].rows,
@@ -172,8 +172,8 @@ impl Network {
 
 #[derive(Clone)]
 pub struct TrainingData {
-    pub inputs: Vec<f32>,
-    pub target: Vec<f32>,
+    pub inputs: Vec<f64>,
+    pub target: Vec<f64>,
     pub classification: u8,
 }
 
@@ -187,12 +187,12 @@ pub struct NetworkData {
 struct WeightData {
     rows: usize,
     cols: usize,
-    data: Vec<f32>,
+    data: Vec<f64>,
 }
 
 #[derive(Serialize, Deserialize)]
 struct BiasData {
     rows: usize,
     // cols is always 1
-    data: Vec<f32>,
+    data: Vec<f64>,
 }
